Change type data of column int to bigint for accommodate big data

+++ po-includes/database/migrations/2019_10_08_114314_create_tags_table.php
@@ -14,12 +14,12 @@ class CreateTagsTable extends Migration
     public function up()
     {
         Schema::create('tags', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 			$table->string('title')->nullable();
 			$table->string('seotitle')->nullable();
-			$table->integer('count')->default('1');
-			$table->integer('created_by')->default('1');
-			$table->integer('updated_by')->default('1');
+			$table->bigInteger('count')->default('1');
+			$table->bigInteger('created_by')->default('1');
+			$table->bigInteger('updated_by')->default('1');
             $table->timestamps();
         });
     }
